+++ ttnn/cpp/ttnn/operations/normalization/layernorm/device/kernels/dataflow/reader_mcast_sender_unary_sharded_ln.cpp
@@ -5,7 +5,7 @@
 #include <stdint.h>
 #include "api/dataflow/dataflow_api.h"
 #include "hostdevcommon/common_values.hpp"
-#include "noc_addr_utils.h"
+#include "layernorm_dataflow_utils.h"
 
 namespace df = norm::layernorm::device::kernels::dataflow;
